# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TRZWorld with Swagger",
        "description": "L5 Swagger OpenApi",
        "contact": {
            "email": "carlosbelchiorsilva@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/exchanges/trade": {
            "post": {
                "tags": [
                    "Trade Items"
                ],
                "summary": "Perform a trade between survivors",
                "operationId": "e03a1d1157d27d2b9d4cb0c89b31bdd6",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "survivor1_id",
                                    "items_to_trade_s1",
                                    "survivor2_id",
                                    "items_to_trade_s2"
                                ],
                                "properties": {
                                    "survivor1_id": {
                                        "description": "ID of the first survivor involved in the trade",
                                        "type": "integer"
                                    },
                                    "items_to_trade_s1": {
                                        "description": "Items to be traded by survivor1",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "item": {
                                                    "description": "ID of the item to be traded by survivor1",
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "description": "Quantity of the item to be traded by survivor1",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "survivor2_id": {
                                        "description": "ID of the second survivor involved in the trade",
                                        "type": "integer"
                                    },
                                    "items_to_trade_s2": {
                                        "description": "Items to be traded by survivor2",
                                        "type": "array",
                                        "items": {
                                            "required": [
                                                "item",
                                                "quantity"
                                            ],
                                            "properties": {
                                                "item": {
                                                    "description": "ID of the item to be traded by survivor2",
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "description": "Quantity of the item to be traded by survivor2",
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating successful trade",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request, invalid parameters provided",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating bad request",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Error message indicating internal server error",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}